refactor(utils): rename `getImage` -> `findImage`
